Move the plan overview table element to the final directory in the file system
