--- html/modules/custom/ghi_blocks/src/Plugin/Block/Home/README.txt
@@ -1,1 +1,0 @@
-The plugins in this directory should be moved to the GlobalPage namespace.
